feat: move icons to tsx components
