+++ packages/website/src/components/docs/navigation/external-links.astro
@@ -2,7 +2,7 @@
 import { BookOpen, Github, Map, MessageCircle } from 'lucide-react'
 import { Flex, Stack } from 'styled-system/jsx'
 import { Link } from '~/components/ui/link'
-import { FrameworkToggleGroup } from './framework-toggle-group'
+import { FrameworkToggleGroup } from './framework-toggle/framework-toggle-group'
 
 const links = [
   {
