+++ packages/website/src/components/docs/navigation/framework-toggle/icons/index.ts
@@ -1,0 +1,3 @@
+export { ReactIcon } from './react'
+export { SolidIcon } from './solid'
+export { VueIcon } from './vue'
