+++ packages/website/src/components/docs/navigation/framework-toggle/icons/react.tsx
@@ -1,0 +1,12 @@
+export const ReactIcon = () => {
+  return (
+    <svg viewBox="0 0 23 21" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M17.7716 6.76796C20.6334 7.63356 22.5 9.02939 22.5 10.5023C22.5 11.9706 20.6426 13.3664 17.7991 14.232C18.4641 17.1053 18.1752 19.3939 16.8865 20.1281C16.5838 20.3003 16.2352 20.3819 15.85 20.3819C14.5888 20.3819 12.9377 19.5027 11.2683 17.9709C9.59896 19.5117 7.94792 20.4 6.68671 20.4C6.30146 20.4 5.95291 20.3184 5.65481 20.1462C4.36149 19.412 4.06798 17.1189 4.72839 14.2366C1.86659 13.371 0 11.9751 0 10.5023C0 9.03392 1.85742 7.64262 4.71005 6.77249C4.04504 3.89925 4.33398 1.61063 5.6227 0.876454C5.92539 0.70424 6.27395 0.622666 6.65919 0.622666V0.618134C7.9204 0.618134 9.57144 1.49733 11.2408 3.02912C12.9102 1.48826 14.5612 0.600006 15.8225 0.600006V0.604542C16.2013 0.605195 16.5483 0.686805 16.8452 0.858326C18.1385 1.5925 18.432 3.88565 17.7716 6.76796ZM15.8179 1.61063C14.864 1.61304 13.4627 2.36513 11.97 3.7361C12.6167 4.40229 13.2588 5.16366 13.8779 6.00206C14.9098 6.11083 15.8912 6.27851 16.7901 6.49604C16.891 6.03832 16.9736 5.58966 17.024 5.16366C17.235 3.43246 16.9644 2.08648 16.3407 1.73299C16.1985 1.65594 16.0288 1.61516 15.8179 1.61516V1.61063ZM10.5116 3.74063C9.01651 2.37652 7.60854 1.62875 6.65919 1.62875V1.63329C6.44822 1.63329 6.26936 1.66954 6.13177 1.75112C5.51263 2.10914 5.24663 3.45512 5.46219 5.18632C5.51722 5.60778 5.59519 6.04738 5.70067 6.50057C6.59957 6.27851 7.58102 6.11083 8.61751 6.00206C9.22747 5.16366 9.86954 4.40682 10.5116 3.74063ZM15.9096 9.26958C16.1481 8.66683 16.359 8.06409 16.5287 7.47947C15.9279 7.33445 15.2996 7.21662 14.6484 7.12145C14.8731 7.46587 15.0978 7.81937 15.3088 8.18192C15.5243 8.54447 15.7216 8.90703 15.9096 9.26958ZM13.323 14.0372C13.7036 13.4979 14.0705 12.9178 14.4282 12.306C14.7722 11.7123 15.0887 11.105 15.3776 10.4977C15.0887 9.89045 14.7722 9.28771 14.4237 8.69403C14.0659 8.08675 13.6944 7.51119 13.3138 6.97189C12.635 6.91751 11.9425 6.88579 11.25 6.88579C10.5529 6.88579 9.86037 6.91751 9.17702 6.97643C8.79637 7.51573 8.42947 8.09581 8.07174 8.70762C7.72778 9.3013 7.40674 9.90858 7.1224 10.5159C7.41133 11.1231 7.72778 11.7259 8.07633 12.3196C8.43406 12.9268 8.80554 13.5024 9.1862 14.0417C9.86496 14.0961 10.5575 14.1278 11.25 14.1278C11.9471 14.1278 12.6396 14.0961 13.323 14.0372ZM16.5425 13.5205C16.3682 12.9314 16.1572 12.3241 15.9096 11.7168C15.7216 12.0794 15.5244 12.4465 15.3134 12.809C15.1024 13.1761 14.8777 13.5341 14.653 13.8831C15.3088 13.7879 15.9417 13.6656 16.5425 13.5205ZM9.9842 15.1022C10.4061 15.6324 10.8326 16.1173 11.2592 16.5524C11.6811 16.1173 12.1076 15.6324 12.5204 15.1022C12.103 15.1248 11.6765 15.1339 11.2454 15.1339C10.8189 15.1339 10.397 15.1203 9.9842 15.1022ZM5.96667 13.5251C6.56746 13.6701 7.19578 13.7879 7.84702 13.8831C7.62229 13.5387 7.40216 13.1852 7.1866 12.8226C6.97105 12.4601 6.77384 12.0975 6.58581 11.7349C6.34733 12.3377 6.13636 12.9404 5.96667 13.5251ZM12.5112 5.90236C12.0893 5.37212 11.6628 4.88721 11.2362 4.45214C10.8143 4.88721 10.3878 5.37212 9.97503 5.90236C10.3924 5.8797 10.8189 5.87064 11.25 5.87064C11.6765 5.87064 12.0984 5.88423 12.5112 5.90236ZM7.18202 8.19551C7.39298 7.82843 7.61771 7.47041 7.84243 7.12145C7.1866 7.21662 6.55371 7.33898 5.95291 7.47947C6.12719 8.06862 6.33815 8.6759 6.58581 9.28318C6.77384 8.92062 6.97105 8.55807 7.18202 8.19551ZM1.01814 10.5023C1.01814 11.2138 2.06839 12.1111 3.69191 12.7954C4.09091 12.9631 4.52201 13.1172 4.97605 13.2577C5.23288 12.374 5.57226 11.454 5.99419 10.5113C5.56767 9.56416 5.2237 8.63964 4.96229 7.75139C4.51743 7.89188 4.08632 8.04143 3.69191 8.20911C2.06839 8.8889 1.01814 9.79075 1.01814 10.5023ZM5.47594 15.8409C5.26498 17.5721 5.53557 18.9181 6.15929 19.2715C6.29688 19.3486 6.47115 19.3894 6.67753 19.3939C7.63147 19.3939 9.03485 18.6416 10.53 17.2684C9.8833 16.6022 9.24123 15.8409 8.62209 15.0025C7.59019 14.8937 6.60874 14.726 5.70984 14.5085C5.60895 14.9662 5.52639 15.4149 5.47594 15.8409ZM16.3682 19.2534C16.9874 18.8954 17.2534 17.5494 17.0378 15.8182C16.9874 15.3967 16.9048 14.9571 16.7993 14.4994C15.9004 14.7215 14.919 14.8892 13.8825 14.9979C13.2725 15.8363 12.6304 16.5932 11.9884 17.2594C13.4835 18.6235 14.8914 19.3712 15.8408 19.3712C16.0518 19.3712 16.2306 19.335 16.3682 19.2534ZM17.5331 13.2531C17.978 13.1127 18.4091 12.9631 18.8035 12.7954C20.427 12.1156 21.4773 11.2138 21.4818 10.5023C21.4818 9.79075 20.4316 8.89343 18.8081 8.20911C18.4045 8.04143 17.9734 7.88734 17.5194 7.74685C17.2625 8.63058 16.9231 9.55056 16.5012 10.4932C16.9277 11.4404 17.2717 12.3649 17.5331 13.2531ZM11.2502 11.7375C11.9405 11.7375 12.5002 11.1834 12.5002 10.5C12.5002 9.81653 11.9405 9.26249 11.2502 9.26249C10.5598 9.26249 10.0002 9.81653 10.0002 10.5C10.0002 11.1834 10.5598 11.7375 11.2502 11.7375Z"
+        fill="#0B8CEA"
+      ></path>
+    </svg>
+  )
+}
