+++ packages/website/src/components/docs/navigation/framework-toggle/icons/solid.tsx
@@ -1,0 +1,26 @@
+export const SolidIcon = () => {
+  return (
+    <svg viewBox="0 0 24 23" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path
+        d="M6.173 11.2876C5.41479 10.8554 4.46286 9.59832 5.19815 7.39808C5.28833 7.12823 5.4223 6.89271 5.59486 6.68924C6.22431 5.56893 7.17163 3.90122 7.79777 2.8603C12.997 -1.28853 20.1459 3.07638 23.0704 5.77744L20.7958 10.3152C14.5567 12.6489 8.44765 11.9359 6.173 11.2876Z"
+        fill="#0B8CEA"
+      ></path>
+      <path
+        d="M5.19815 7.39808C4.46286 9.59832 5.41479 10.8554 6.173 11.2876C8.44765 11.9359 14.5567 12.6489 20.7958 10.3152M5.19815 7.39808C6.49797 3.50858 16.8964 6.74977 20.7958 10.3152M5.19815 7.39808C5.73974 6.4257 7.01789 4.15681 7.79777 2.8603C12.997 -1.28853 20.1459 3.07638 23.0704 5.77744L20.7958 10.3152"
+        stroke="black"
+        strokeWidth="1.10318"
+      ></path>
+      <path
+        d="M4.21987 12.5857L1.2998 17.4579C11.1631 24.7338 16.2246 21.3558 17.5224 18.7572L20.118 14.5346C20.118 13.1271 19.1446 10.182 15.2512 9.66234C11.3578 9.14263 6.27472 11.3947 4.21987 12.5857Z"
+        fill="#0B8CEA"
+        stroke="black"
+        strokeWidth="1.10318"
+      ></path>
+      <path
+        d="M1.62451 17.4524C7.25734 15.0913 17.5481 11.6569 17.2231 19.3842"
+        stroke="black"
+        strokeWidth="1.10318"
+      ></path>
+    </svg>
+  )
+}
