+++ packages/website/src/components/docs/navigation/framework-toggle/icons/vue.tsx
@@ -1,0 +1,15 @@
+export const VueIcon = () => {
+  return (
+    <svg viewBox="0 0 21 19" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path
+        d="M12.6863 0.950014L10.4 4.89904L8.1137 0.950014H0.5L10.4 18.05L20.3 0.950014H12.6863Z"
+        fill="#41B883"
+      ></path>
+      <path
+        d="M12.6863 0.950015L10.4 4.91922L8.11366 0.950015H4.45996L10.4 11.2622L16.34 0.950015H12.6863Z"
+        fill="white"
+        fillOpacity="0.53"
+      ></path>
+    </svg>
+  )
+}
